Split Mine Dot director code for lint
